Avoid DirectoryNotFoundException when no symbols are present for a package that is being removed.

+++ source/NuGet.Lucene.Web/Symbols/SymbolSource.cs
@@ -30,7 +30,10 @@ namespace NuGet.Lucene.Web.Symbols
         {
             var path = GetPackageSymbolPath(package);
 
-            Directory.Delete(path, recursive:true);
+            if (Directory.Exists(path))
+            {
+                Directory.Delete(path, recursive: true);
+            }
 
             return Task.FromResult(true);
         }
